Multiplayer: thread diversi per invio info ai client

+++ MioSpaceInvaders/src/gui/states/drawer/Drawable.java
@@ -3,5 +3,5 @@ package gui.states.drawer;
 import org.newdawn.slick.Image;
 
 public interface Drawable {
-    void render(Image image, double x, double y, double width, double height);
+    void render(Image image, float x, float y, float width, float height);
 }
