+++ MioSpaceInvaders/src/gui/states/drawer/SpriteDrawer.java
@@ -5,8 +5,8 @@ import org.newdawn.slick.Image;
 
 public class SpriteDrawer implements Drawable {
 
-    public void render(Image image, double x, double y, double width, double height){
-        image.draw((float) (x*SpaceInvaders.SCALE_X),(float) (y*SpaceInvaders.SCALE_Y),
-                (float) (width*SpaceInvaders.SCALE_X), (float) (height*SpaceInvaders.SCALE_Y));
+    public void render(Image image, float x, float y, float width, float height){
+        image.draw((x*SpaceInvaders.SCALE_X),(y*SpaceInvaders.SCALE_Y), (width*SpaceInvaders.SCALE_X),
+                (height*SpaceInvaders.SCALE_Y));
     }
 }
