+++ MioSpaceInvaders/src/logic/sprite/Coordinate.java
@@ -2,26 +2,26 @@ package logic.sprite;
 
 public class Coordinate {
 
-    private double x, y;
+    private float x, y;
 
-    public Coordinate(double x, double y) {
+    public Coordinate(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
