+++ MioSpaceInvaders/src/logic/sprite/unmovable/Brick.java
@@ -6,7 +6,7 @@ import logic.sprite.Sprite;
 public class Brick extends Sprite {
     private int life;
 
-    public Brick(Coordinate coordinate, double width, double height) {
+    public Brick(Coordinate coordinate, float width, float height) {
         super(coordinate, width, height);
         life = 4;
     }
